changed the signatues of the builtin functions
now they receive a node instead of t_databus (this is possible because
with the getter_data() function, anyone can get the t_databus from
anywhere always, and this will be usefull because once we're dealing
with pipes, the node where the builtin command actually is will not be
simply data->cmds->head, it will be a specific node for each situation.
So, changing the signature allows us to pass the specific node for each.

	modified:   Makefile
	modified:   include/minishell.h
	new file:   lib/src/utils/ft_atoi
	modified:   src/builtins/echo_builtin.c
	modified:   src/builtins/env_builtin.c
	modified:   src/builtins/exit_builtin.c
	modified:   src/builtins/export_builtin.c
	modified:   src/builtins/unset_builtin.c
	renamed:    src/builtins/utils_builtins.c -> src/builtins/utils_builtins_01.c
	new file:   src/builtins/utils_builtins_02.c
	modified:   src/executor/executor.c